requirements.txt
MarkupSafe~=2.1.4
werkzeug~=3.0.1
numpy~=1.24.4
scipy~=1.10.1
joblib~=1.3.2
pandas~=2.0.3
flask~=3.0.1
scikit-learn~=1.1.2
pyjwt~=2.8.0
python-dotenv~=1.0.1
Werkzeug==3.0.1
Gunicorn
flask-sqlalchemy
requests~=2.31.0